add FormattedMessage interface to Translation component
